simplification of projection matrix
the original one was based on some old opengl code from another
project of mine.  new one is much more basic and as a result the
translation controls are no longer backwards.  i might do some sort of
transformation on the z-axis in the future because it seems very
out-of-scale compared to x and y but other than that it's fine.